fix:split eventCard
